+++ Cap8_Exercicios/PEOO_Cap8_Exercicio8.4.8.py
@@ -4,7 +4,7 @@
 # Autor...: Givanaldo Rocha de Souza
 # Data....: 02/05/2020
 
-from contas import ContaInvestimento
+from classes import ContaInvestimento
 
 conta = ContaInvestimento(1000.0, 10)
 print("Saldo inicial: R$ %.2f" % conta.saldo)
