--- Cap8_Exercicios/questao_01/ingresso.py
@@ -1,16 +1,0 @@
-# Livro...: Introdução a Python com Aplicações de Sistemas Operacionais
-# Capítulo: 08
-# Questão.: 8.4.1
-# Autor...: Givanaldo Rocha de Souza
-# Data....: 30/04/2020
-
-class Ingresso:
-    def __init__(self, evento, valor):
-        self.evento = evento
-        self.valor = valor
-
-    def exibirValor(self):
-        return self.valor
-
-    def __str__(self):
-        return ("%s: R$ %.2f" % (self.evento, self.valor))
